Rename: pages 내 폴더명 파스칼 케이스로 변경
