+++ src/pages/DetailInfo/DetailInfo.tsx
