+++ src/pages/DetailInfo/DetailInfoStyle.tsx
