+++ src/pages/DetailInfo/ImageSlider.tsx
